+++ packages/expo-cli/src/commands/eject/updatePackageJson.ts
@@ -194,13 +194,13 @@ function updatePackageJSONScripts({ pkg }: { pkg: PackageJSONConfig }) {
     pkg.scripts = {};
   }
   if (!pkg.scripts.start?.includes('--dev-client')) {
-    pkg.scripts.start = 'react-native start';
+    pkg.scripts.start = 'expo start --dev-client';
   }
   if (!pkg.scripts.android?.includes('run')) {
-    pkg.scripts.android = 'react-native run-android';
+    pkg.scripts.android = 'expo run:android';
   }
   if (!pkg.scripts.ios?.includes('run')) {
-    pkg.scripts.ios = 'react-native run-ios';
+    pkg.scripts.ios = 'expo run:ios';
   }
 }