+++ packages/expo-cli/src/commands/start/TerminalUI.ts
@@ -50,9 +50,9 @@ const div = chalk.dim(`│`);
 export async function shouldOpenDevToolsOnStartupAsync() {
   return UserSettings.getAsync(
     'openDevToolsAtStartup',
-    // Defaults to true for new users.
-    // TODO: switch this to false.
-    true
+    // Defaults to false for new users.
+    // We can swap this back to true when dev tools UI has a code owner again.
+    false
   );
 }
 
@@ -181,10 +181,10 @@ const printServerInfo = async (
       urlOpts.printQRCode(url);
       Log.nested(item(`Metro waiting on ${u(url)}`));
       // Log.newLine();
-      // TODO: if dev client, change this message!
+      // TODO: if development build, change this message!
       Log.nested(item(`Scan the QR code above with Expo Go (Android) or the Camera app (iOS)`));
     } catch (error) {
-      // @ts-ignore: If there is no dev client scheme, then skip the QR code.
+      // @ts-ignore: If there is no development build scheme, then skip the QR code.
       if (error.code !== 'NO_DEV_CLIENT_SCHEME') {
         throw error;
       } else {