+++ packages/prebuild-config/src/plugins/unversioned/react-native-maps.ts
@@ -1,4 +1,4 @@
-import { AndroidConfig, ConfigPlugin, IOSConfig } from '@expo/config-plugins';
+import { AndroidConfig, ConfigPlugin, IOSConfig, withInfoPlist } from '@expo/config-plugins';
 import resolveFrom from 'resolve-from';
 
 import { createLegacyPlugin } from './createLegacyPlugin';
@@ -8,15 +8,19 @@ const LOCATION_USAGE = 'Allow $(PRODUCT_NAME) to access your location';
 // Copied from expo-location package, this gets used when the
 // user has react-native-maps installed but not expo-location.
 const withDefaultLocationPermissions: ConfigPlugin = config => {
+  // TODO: Autolinking
+  const isLinked = true;
   // Only add location permissions if react-native-maps is installed.
   if (
     config._internal?.projectRoot &&
-    resolveFrom.silent(config._internal.projectRoot, 'react-native-maps')
+    resolveFrom.silent(config._internal.projectRoot, 'react-native-maps') &&
+    isLinked
   ) {
-    if (!config.ios) config.ios = {};
-    if (!config.ios.infoPlist) config.ios.infoPlist = {};
-    config.ios.infoPlist.NSLocationWhenInUseUsageDescription =
-      config.ios.infoPlist.NSLocationWhenInUseUsageDescription || LOCATION_USAGE;
+    config = withInfoPlist(config, config => {
+      config.modResults.NSLocationWhenInUseUsageDescription =
+        config.modResults.NSLocationWhenInUseUsageDescription || LOCATION_USAGE;
+      return config;
+    });
 
     return AndroidConfig.Permissions.withPermissions(config, [
       'android.permission.ACCESS_COARSE_LOCATION',