+++ packages/xdl/src/start/startDevServerAsync.ts
@@ -11,6 +11,7 @@ import {
   assertValidProjectRoot,
   ExpoUpdatesManifestHandler,
   getFreePortAsync,
+  LoadingPageHandler,
   ManifestHandler,
   ProjectSettings,
   ProjectUtils,
@@ -66,9 +67,14 @@ export async function startDevServerAsync(
     options.maxWorkers = startOptions.maxWorkers;
   }
 
-  const { server, middleware, messageSocket } = await runMetroDevServerAsync(projectRoot, options);
   // TODO: reduce getConfig calls
   const projectConfig = getConfig(projectRoot, { skipSDKVersionRequirement: true });
+
+  // Use the unversioned metro config.
+  options.unversioned =
+    !projectConfig.exp.sdkVersion || projectConfig.exp.sdkVersion === 'UNVERSIONED';
+
+  const { server, middleware, messageSocket } = await runMetroDevServerAsync(projectRoot, options);
 
   const easProjectId = projectConfig.exp.extra?.eas?.projectId;
   const useExpoUpdatesManifest =
@@ -87,6 +93,8 @@ export async function startDevServerAsync(
       ? ExpoUpdatesManifestHandler.getManifestHandler(projectRoot)
       : ManifestHandler.getManifestHandler(projectRoot)
   );
+
+  middleware.use(LoadingPageHandler.getLoadingPageHandler(projectRoot));
 
   return [server, middleware, messageSocket];
 }